requirements.txt
fastapi[all]
uvicorn
sqlmodel
psycopg2-binary
python-dotenv